Altered pdftohtml to output em-based units.
Running "pdftohtml file.pdf" with no other arguments will create
file.html, containing HTML divs with the minimum necessary CSS
styling information to position the text with HTML em units.
This allows browser font-resizing to work correctly.
No font information is included in the HTML file.

The output is not a complete HTML file, but rather a fragment
of UTF-8 encoded HTML to be included in an HTML template.

This is a messy hack, and breaks some command line arguments,
notably "-xml".  But "-stdout" still works.